+++ src/kademlia/utils/UInt128.h
@@ -56,42 +56,42 @@ public:
 
 	CUInt128();
 	CUInt128(bool fill);
-	CUInt128(ULONG value);
+	CUInt128(uint32 value);
 	CUInt128(const byte *valueBE);
 	/**
 	 * Generates a new number, copying the most significant 'numBits' bits from 'value'.
 	 * The remaining bits are randomly generated.
 	 */
-	CUInt128(const CUInt128 &value, UINT numBits = 128);
+	CUInt128(const CUInt128 &value, uint32 numBits = 128);
 
 	const byte* getData() const { return (byte*)m_data; }
 	byte* getDataPtr() const { return (byte*)m_data; }
 
 	/** Bit at level 0 being most significant. */
-	UINT getBitNumber(UINT bit) const;
+	uint32 getBitNumber(uint32 bit) const;
 	int compareTo(const CUInt128 &other) const;
-	int compareTo(ULONG value) const;
+	int compareTo(uint32 value) const;
 
 	void toHexString(wxString *str) const;
 	void toBinaryString(wxString *str, bool trim = false) const;
 	void toByteArray(byte *b) const;
 
-	ULONG get32BitChunk(int val) const {return m_data[val];}
+	uint32 get32BitChunk(int val) const {return m_data[val];}
 
 	CUInt128& setValue(const CUInt128 &value);
-	CUInt128& setValue(ULONG value);
+	CUInt128& setValue(uint32 value);
 	CUInt128& setValueBE(const byte *valueBE);
 
 	CUInt128& setValueRandom(void);
 //	CUInt128& setValueGUID(void);
 
-	CUInt128& setBitNumber(UINT bit, UINT value);
-	CUInt128& shiftLeft(UINT bits);
+	CUInt128& setBitNumber(uint32 bit, uint32 value);
+	CUInt128& shiftLeft(uint32 bits);
 
 	CUInt128& add(const CUInt128 &value);
-	CUInt128& add(ULONG value);
+	CUInt128& add(uint32 value);
 	CUInt128& subtract(const CUInt128 &value);
-	CUInt128& subtract(ULONG value);
+	CUInt128& subtract(uint32 value);
 
 	CUInt128& xorv(const CUInt128 &value);
 	CUInt128& xorBE(const byte *valueBE);
@@ -106,19 +106,19 @@ public:
 	bool operator== (const CUInt128 &value) const {return (compareTo(value) == 0);}
 	bool operator!= (const CUInt128 &value) const {return (compareTo(value) != 0);}
 
-	void operator+  (ULONG value) {add(value);}
-	void operator-  (ULONG value) {subtract(value);}
-	void operator=  (ULONG value) {setValue(value);}
-	bool operator<  (ULONG value) const {return (compareTo(value) <  0);}
-	bool operator>  (ULONG value) const {return (compareTo(value) >  0);}
-	bool operator<= (ULONG value) const {return (compareTo(value) <= 0);}
-	bool operator>= (ULONG value) const {return (compareTo(value) >= 0);}
-	bool operator== (ULONG value) const {return (compareTo(value) == 0);}
-	bool operator!= (ULONG value) const {return (compareTo(value) != 0);}
+	void operator+  (uint32 value) {add(value);}
+	void operator-  (uint32 value) {subtract(value);}
+	void operator=  (uint32 value) {setValue(value);}
+	bool operator<  (uint32 value) const {return (compareTo(value) <  0);}
+	bool operator>  (uint32 value) const {return (compareTo(value) >  0);}
+	bool operator<= (uint32 value) const {return (compareTo(value) <= 0);}
+	bool operator>= (uint32 value) const {return (compareTo(value) >= 0);}
+	bool operator== (uint32 value) const {return (compareTo(value) == 0);}
+	bool operator!= (uint32 value) const {return (compareTo(value) != 0);}
 
 private:
 
-	ULONG m_data[4];
+	uint32 m_data[4];
 };
 
 } // End namespace
